DEV-60066: Fix the API breakage caused by DEV-57956.
In DEV-57956, we applied Jackson annotations to all the standard styles,
but this had the side effect of defaulting forceJacksonPropertyNames to
true, and so all the properties on the generated immutables were forced
to be named according to immutables principles. This broke compatibility
with the pre-225 codebase, and also causes problems now for the Book
with Google project who want to use SNAKE_CASE, but are prevented from
doing so with the standard styles by this property default.